requirements.txt
liac-arff==2.5.0
matplotlib==3.4.2
numpy==1.20.3
scikit-learn==0.24.2
scipy==1.6.2